backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-dotenv
requests
dnspython
sqlalchemy